Cycles-Bake: 1st round of code-review
For the individual commits:
https://github.com/dfelinto/blender-git/tree/bake-cycles-0.1

Documentation:
http://wiki.blender.org/index.php/User:Dfelinto/CyclesBaking

Open Issues:
https://github.com/dfelinto/blender-git/issues?labels=bake-cycles&page=1&state=open

Supported Features:
------------------
* Margin
* User Interface (no more need to fiddle with Python scripts)
* Selected to Active
* Cage Extrusion
* Custom Cage
* File Formats (save as PNG, JPG, OpenEXR, Tiff, Targa and BMP)
* Color Depth/Mode: RGB/RGBA - 8/16/32
* Normal swizzle (change the axis that gets mapped to RGB)
* Normal space (save as tangent, world, object)

Supported Passes:
-----------------
Data Passes
* Normal
* UV
* Diffuse/Glossy/Transmission/Subsurface/Emit Color

Light Passes
* AO
* Combined
* Shadow
* Diffuse/Glossy/Transmission/Subsurface/Emit Direct/Indirect
* Environment

Implementation Notes:
---------------------
Baking for Cycles was implemented in two fronts:

In Blender we created a new operator that will try to bake using the current engine.
So far only Cycles has this implemented, but other engines (even the Blender Internal) could go through this API/pipeline.
The operator is responsible for the image handling, normal space
transformation, populate the UVs, loop over faces, ...

In Cycles we extended the baking functionality available for environment
and displacement for the other render passes (not all of them since some
of them doesn't make sense for baking).

The API is simple. Blender send a populated array of BakePixels to the
renderer, and gets back an array of floats with the result.

BakePixel is tjhe following struct:

struct BakePixel {
    int primitive_id;
    float u, v;
    float dudx, dudy;
    float dvdx, dvdy;
};