Move "Output File" section to the bottom of the bake panel.
The reason for this is heirarchy. During the order of operations the output file is likely one of the last
things you'll set. First the type and width, then margin, then specifc settings, and finally the output file.

Doing this also helps avoid breaking up the settings that are specific to baking, such as the Type and "Selected to Active".